Massive indexing overhaul.
The biggest changes to indexing are:

  1. Diffs larger than 10MB are not indexed.
  2. Commits are indexed before files.
  3. InputDocumentQueue has been removed in favor of the ConcurrentUpdateSolrServer.
  4. Commits are only ever indexed once, across all refs
  5. Refs are currently not indexed with commits.

The last change is a result of the way Hakea walks over each commit in a repository. To avoid indexing the same commit more than once Hakea now adds every ref to the RevWalk. Unfortunately this means that Hakea is temporarily unable to determine what refs apply to particular commits.