Move the processors and Solr document dispatcher to the core library.
The goal is to keep the core functionality of Hakea separate from any implementations. Hakea indexers can now be built as anything from a command line utility to the provided Dropwizard service.

Unfortunately, due to how Hakea is configured, the core library still depends on Dropwizard. In the future the concerns of configuration will be moved out of the code library and into any implementations.